+++ resources/views/components/table.blade.php
@@ -1,8 +1,8 @@
-<div class="overflow-x-auto relative">
+<div class="overflow-x-auto relative rounded-lg shadow-lg">
     <div class="align-middle inline-block min-w-full">
-        <div class="overflow-x-auto relative rounded-lg shadow-lg">
+        <div class="overflow-x-auto relative">
             <table
-                class="text-sm text-left text-gray-500 dark:text-gray-300 min-w-full divide-y divide-gray-200 dark:divide-slate-700 shadow-lg">
+                class="text-sm text-left text-gray-500 dark:text-gray-300 min-w-full divide-y divide-gray-200 dark:divide-slate-700">
                 <thead
                     class="text-xs text-slate-800 uppercase bg-slate-200 dark:bg-gray-700 dark:text-gray-400 select-none">
                     <tr>