+++ src/Components/Table/Heading.php
@@ -7,19 +7,21 @@ use Illuminate\View\Component;
 class Heading extends Component
 {
     public $mobileOnly;
-
     public $fullWidth;
+    public $sortable;
+    public $direction;
 
     /**
      * Create a new component instance.
      *
      * @return void
      */
-    public function __construct($mobileOnly = '', $fullWidth = '')
+    public function __construct($mobileOnly = null, $fullWidth = null, $sortable = null, $direction = null)
     {
         $this->mobileOnly = $mobileOnly;
-
         $this->fullWidth = $fullWidth;
+        $this->sortable = $sortable;
+        $this->direction = $direction;
     }
 
     /**
@@ -29,6 +31,6 @@ class Heading extends Component
      */
     public function render()
     {
-        return view('wireblade::components.table.heading');
+        return view("wireblade::components.table.heading");
     }
 }